Revert "refactor auth and user"
This reverts commit 22bf82e42c3e705aad4dbc3b337686f9c0d669cf.
