+++ src/entities/user/api/types.ts
@@ -1,4 +1,4 @@
-import type { User } from "@/entities/user/model/types";
+import type { User } from "../model/types";
 
 export interface OAuthExchangeParams {
   code: string;
