+++ src/entities/user/index.ts
@@ -1,5 +1,7 @@
-export { userKeys } from './model/keys'
+export { queryKeys as userKeys } from './api/queryKeys'
 export { userRequests } from './api/requests'
-export { useMe } from './api/queries'
+export { type AuthResponse, type OAuthExchangeParams } from './api/types'
 export { type User } from './model/types'
-export { UserSlot } from "./ui/UserSlot/UserSlot"
+export { useMe } from './lib/hooks/useMe'
+export { useAuthStore } from './model/store/useAuthStore'
+export {UserSlot} from "./ui/UserSlot/UserSlot"
