--- src/entities/user/model/keys.ts
@@ -1,5 +1,0 @@
-export const userKeys = {
-  all: ['users'] as const,
-  me: () => [...userKeys.all, 'me'] as const,
-  user: (id: string) => [...userKeys.all, id] as const,
-};
