+++ src/features/auth/index.ts
@@ -1,2 +1,1 @@
 export { useLogin, useLogout } from './api/mutations'
-export {useAuthStore} from './store/useAuthStore'
