+++ tsconfig.app.json
@@ -25,8 +25,9 @@
     "noUncheckedSideEffectImports": true,
 
     /* PATHS */
+    "baseUrl": ".",
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": ["src/*"]
     }
   },
   "include": ["src", "vite-env.d.ts"]
